Fallback to body->size if payLoadsize is empty
For some reason body->size is exactly as big as the
payload I'm trying to send (using android app). This is
probably a bug in the Android app, I decided to commit
the workaround anyway so we keep compatibility with
current (and old) versions of the Android app.

CCMAIL: albertvaka@gmail.com