Add an alternative way of executing kdeconnectd via autostart
We are currently using KDED only to start the daemon which makes little
sense. The reason for this was so we can start/stop the service at will
via KDED startup interface.

This sucks, it is a horrible interface for starting/disabling kdeconnect
and makes us depend on KDED for no good reason.

So what this patch does is make use of XDG autostart which is supported
in all major desktops (GNOME/UNITY/XFACE/KDE...).

KDE has an interface to disable autostarted programs (although the
interface sucks).