Set a valid url for UDS_URL, basically one that includes files://
In Qt5 QUrl need to be created by calling fromLocalFile to handle
paths such /root/foo.txt.
So KIO apparently broke compatibility with setting such paths in
UDS_URL and we have to pass instead files:///root/foo.txt

Altought this might be a bug in KIO, we want to work with all KF5 based
versions of plasma and this patch will just do that.

REVIEW:124544