Remove the keys from the notification hashmap
When disabling the notifications plugin, we can't delete the notification
adapter due to a QT bug. Thus we leak it, and to not expose the notifications
anymore, we clean up the notifications in the adaptor. Therefore we delete
the QStrings in the notification hashmap (mNotifications) with qDeleteAll().
But this function does only free the QStrings, but no remove the keys from the
hashmap. Thus, a subsequent call to activeNotifications() has an undefined
behavior, as there are dangling pointers in the hashmap. This is solved by
removing the keys from the hashmap.

REVIEW: 127543