Delay kdeconnectd autostart phase
Summary:
This value maps to KAutostart::StartPhase

kdeconnectd is certainly not BaseDesktop which is "essential desktop
services" which is at the same level as plasmashell.

This delays it into DesktopServices which is slightly later.

For a user this won't change a thing, it'll still be loaded before
ksplash is removed.

Test Plan: None

Differential Revision: https://phabricator.kde.org/D9031