Fix another DBus connection in Kirigami app
Summary: When a DevicePage is created from FindDevicePage the device is not passed correctly, resulting in some device is null warnings. The pair button still does not work, but it seems to be another issue.

Test Plan: I could set the button text to the device id, so the device is passed correctly. The button still does not work as intended though.

Reviewers: #kde_connect

Subscribers: apol, #kde_connect

Tags: #kde_connect

Differential Revision: https://phabricator.kde.org/D12506