Enable desktop2desktop keyboard input
Summary:
Fixes T6701

The reason it did not work before is that the "Keyboard" did not report itself as available.

Test Plan:
Type in plasmoid on one Desktop, see text on another.
Note that no ack packets are sent, so the text does not appear in the original input text field. I want to await the discussion in D12670 before doing that

Reviewers: apol

Reviewed By: apol

Subscribers: mtijink, kdeconnect

Tags: #kde_connect

Maniphest Tasks: T6701

Differential Revision: https://phabricator.kde.org/D12812