Fix include guards
Summary:
Some plugins were copy/pasted from a template and forgot to change their include guards. Since plugin header files are never cross-referenced, this was not a serious issue, but it looks nicer to be correct

NotificationListener.h did not have an include guard. As before, this is not a problem currently, but it's best to have it fixed

Test Plan: Project should build, compile, and run as normal

Reviewers: #kde_connect, nicolasfella

Reviewed By: #kde_connect, nicolasfella

Subscribers: kdeconnect

Tags: #kde_connect

Differential Revision: https://phabricator.kde.org/D16437