Added new indicator icons for dark and clear backgrounds
Default to dark because it's more common (Windows, Gnome) but we need to
implement a way to chose automatically between the dark and clear icon
(and this is gonna be platform-specific).

Note that dashes in names have a special meaning when loading icons. The
new icon file names are ugly for this reason.