Define content margins for settings app
The settings app is supposed to be used on its own. The buttons are without
any margins to the window border though (when borders are deactivated - the
default on Breeze).

Set the default layout margins to improve the look.

(cherry picked from commit f953678e2e1ed39512dcdb8861772d076b04195b)