Work around SMS not being sent from cli
Something in the demarshalling of the address list is broken and I can't figure out why.
Instead expose a much simpler method that only takes one address string.
This removes the ability to send to multiple destinations from one command, but that's better than not working at all.

This patch is for 20.08 only as the marshalling works differently in master anyway.