[SMS App] Prevent crash if a conversation doesn't have any messages
## Summary

Fix crash in kdeconnect-sms if a conversation contains a message which has no addresses.

I am not sure why this would happen and there's probably something which should be fixed on the Android side to ensure that every message has addresses, but this is an easy fix to prevent user-reported crashes.

BUG: 449719

## Test Plan

### Before:
kdeconnect-sms crashes upon selecting a conversation with a message with no addresses

### After:
kdeconnect-sms does not crash